Make the standalone package work again
